+++ src/app.js
@@ -1,6 +1,8 @@
 import Fastify from 'fastify';
-import { getDB, saveDB } from './utils/db/index.js';
-import { v4 } from 'uuid';
+import { general } from './services/general/index.js';
+import { createTodo } from './services/todos/create-todo.js';
+// import { getManyTodo } from './services/todos/get-many-todo.js';
+// import { getTodo } from './services/todos/get-todo.js';
 
 const prefix = '/api';
 
@@ -8,35 +10,16 @@ export async function build () {
   // initialize fastify
   const fastify = Fastify({ logger: true });
 
-  fastify.get(prefix, async (request, reply) => {
-    return { success: true };
-  });
+  fastify.get(prefix, general);
 
   // create todo
-  fastify.post(`${prefix}/todo`, async (request, reply) => {
-    const { body } = request;
-    const { title, description, isDone = false } = body;
-    const db = await getDB();
+  fastify.post(`${prefix}/todo`, createTodo);
 
-    const id = v4();
+  // get many todo
+  // fastify.get(`${prefix}/todo`, getManyTodo);
 
-    const todo = {
-      title,
-      description,
-      isDone,
-      createdDate: new Date().getTime(),
-      updatedDate: new Date().getTime()
-    };
-
-    db.todos[id] = todo;
-
-    await saveDB(db);
-
-    return {
-      id,
-      ...todo
-    };
-  });
+  // get one todo
+  // fastify.get(`${prefix}/todo/:todoId`, getTodo);
 
   return fastify;
 }
